TC1/Docker/MariaDBPackage/MariaDB-Memcached/app/requirements.txt
flask
pymemcache
mysql-connector-python
